[Inference]ADD Bench Chatglm2 script (#4963)
* add bench chatglm

* fix bug and make utils

---------
